refactor: extract getDisplayTextForTemplate into shared template-utils
Removes duplicated getDisplayTextForTemplate from template-badge-input
and template-badge-textarea. Both now import from keeperhub/lib/template-utils
alongside the existing doesNodeExist.
