+++ keeperhub/lib/template-utils.ts
@@ -1,6 +1,17 @@
+import { findActionById } from "@/plugins";
 import { BUILTIN_NODE_ID } from "./builtin-variables";
 
 const TEMPLATE_REF_PATTERN = /\{\{@([^:]+):([^}]+)\}\}/;
+
+/** Minimal node shape needed by template helpers */
+export type TemplateNode = {
+  id: string;
+  data: {
+    label?: string;
+    type?: string;
+    config?: Record<string, unknown>;
+  };
+};
 
 /**
  * Checks whether a template reference like {{@nodeId:Label.field}}
@@ -21,4 +32,49 @@ export function doesNodeExist(
     return true;
   }
   return nodes.some((n) => n.id === nodeId);
+}
+
+/**
+ * Gets display text for a template badge by looking up the current node label.
+ * Resolves {{@nodeId:OldLabel.field}} to "CurrentLabel.field".
+ */
+export function getDisplayTextForTemplate(
+  template: string,
+  nodes: readonly TemplateNode[]
+): string {
+  const match = template.match(TEMPLATE_REF_PATTERN);
+  if (!match) {
+    return template;
+  }
+
+  const nodeId = match[1];
+  const rest = match[2];
+
+  const node = nodes.find((n) => n.id === nodeId);
+  if (!node) {
+    return rest;
+  }
+
+  let displayLabel: string | undefined = node.data.label;
+  if (!displayLabel && node.data.type === "action") {
+    const actionType = node.data.config?.actionType as string | undefined;
+    if (actionType) {
+      const action = findActionById(actionType);
+      displayLabel = action?.label;
+    }
+  }
+
+  const dotIndex = rest.indexOf(".");
+
+  if (dotIndex === -1) {
+    return displayLabel ?? rest;
+  }
+
+  const field = rest.substring(dotIndex + 1);
+
+  if (!displayLabel) {
+    return rest;
+  }
+
+  return `${displayLabel}.${field}`;
 }
